Refactored tabs management
